Rename GSYM interfaces and V1 classes for consistency
- Rename interfaces: GsymReaderBase → GsymReader, GsymCreatorBase → GsymCreator
- Rename V1 classes: GsymReader → GsymReaderV1, GsymCreator → GsymCreatorV1
- Move V1 classes to new files: GsymReaderV1.h/.cpp, GsymCreatorV1.h/.cpp
- GsymReader.h and GsymCreator.h now contain only the abstract interfaces
- Update all consumers, tests, and forward declarations

User prompt:
"For the gsym v2 (64-bit) project, sorry about the back and forth, but
can you rename the interface/class and file names, so that:
1. The interfaces should be named GsymReader and GsymCreator.
2. The classes should be named with V1/V2 suffixes (the V2 classes
   already do, so you only need to change the V1 classes, which
   currently don't have the V1 suffix).
3. Put the interfaces and v1/v2 classes into their corresponding file
   names (i.e. create GsymReaderV1.h/cpp and GsymCreatorV1.h/cpp)."
